requirements.txt
ebooklib==0.18
beautifulsoup4==4.12.3
lxml==5.3.0
chromadb==0.6.3
sentence-transformers==3.4.1
openai>=1.0.0
fastapi==0.115.12
uvicorn==0.34.0
python-multipart==0.0.20
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-core>=0.3.0
tavily-python>=0.5.0
